Microservices/analytics-service/requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
gunicorn==23.0.0
python-dotenv==1.0.1
boto3==1.35.35